Fix Priority 2 code review issues: ISBN normalization (strip separators), @Max on pageSize, LIKE wildcard escaping, COUNT(*) mapped to Long instead of Integer (#50)

+++ backend/src/main/java/pl/jaro/restapiworkshop/repository/BookSearchRepository.java
@@ -8,5 +8,5 @@ public interface BookSearchRepository {
 
     Collection<Book> searchBooks(String search, int page, int pageSize, Long userId);
 
-    int countBySearch(String search, Long userId);
+    long countBySearch(String search, Long userId);
 }
